% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{blanks_to_nas}
\alias{blanks_to_nas}
\title{Convert factor levels == "" to NA}
\usage{
blanks_to_nas(x)
}
\arguments{
\item{x}{A factor vector to relevel in the specified way.}
}
\description{
It's probably better to directly use the na.strings argument in
read.csv or fread
}
\examples{
x = factor(c("A", "A", "B", "", "A", "B"))
blanks_to_nas(x)
}

