% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{cast_wrapper}
\alias{cast_wrapper}
\title{Wrapper for casting tables from long to wide}
\usage{
cast_wrapper(X, row_var, col_var, value_var = NULL, aggregate_fun = "sum")
}
\arguments{
\item{X}{The data set to cast}

\item{row_var}{The variable whose unique values will determine the rows.}

\item{col_var}{The variable whose unique values will determine the columns.}

\item{value_var}{The variable to use to fill in the cells of the wide matrix.}

\item{aggregate_fun}{The name of a function to use to summarize many elements
with the same row_var / col_var identification. Must be one of}
}
\value{
The wide version of X.
}

