% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{convert_classes}
\alias{convert_classes}
\title{Convert the columns in a data.table to specified types}
\usage{
convert_classes(X, classes_df, date_formats = NULL)
}
\arguments{
\item{X}{The data.frame or data.table whose columns we need to change classes
of.}

\item{classes_df}{A data.frame with two columns, "name" and "class",
specifying the desired class for each column name in X.}

\item{date_formats}{A vector of date formatting types to try when convert
dates to POSIXlt class.}
}
\value{
X The version of X with classes as specified in classes_df.
}
\description{
Given a data.frame with incorrect types, convert them to
the types in a prespecified classes_df data frame.
}
\examples{
classes_df <- data.frame(name = c("A", "B", "C"), class = c("numeric", "factor", "character"))
X <- data.table(cbind(A = c("1", "121", "31", "132"), B = c("U", "U", "V", "W"), C = c(14, 12, 1, 1))) # starts all character
convert_classes(X, classes_df)
}

