% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_convert_fun}
\alias{get_convert_fun}
\title{Functional to get type conversion functions}
\usage{
get_convert_fun(type, date_formats = NULL)
}
\arguments{
\item{type}{A string specifying the function to retrieve}

\item{date_formats}{A vector of date formatting types to try when convert
dates to POSIXlt class. If none provided, defaults to
c("%Y-%m-%d %H:%M:%S", "%m/%d/%Y", "%Y-%m-%d")}
}
\value{
convert_fun A function that can be used to convert types.
}
\examples{
get_convert_fun("numeric")
}

