% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurizing.R
\name{get_date_features}
\alias{get_date_features}
\title{Get features associated with a POSIXct date}
\usage{
get_date_features(x)
}
\examples{
Sys.Date() \%>\%
  as.POSIXct() \%>\%
  get_date_features()
}

