% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{levels_to_int}
\alias{levels_to_int}
\title{Convert long factor names to integers}
\usage{
levels_to_int(X)
}
\arguments{
\item{X}{The data.table or data.frame whose factors we want to convert
to integer factors.}
}
\value{
A list containing the following elements, \cr
   $X The original data, but with factor levels recoded to integer factors. \cr
   $levels_code A data.frame describing the mapping between original factor
    levels and new integer factor levels.
}

