% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_constant_cols}
\alias{remove_constant_cols}
\title{Remove columns that all one number / level / all NA}
\usage{
remove_constant_cols(X)
}
\arguments{
\item{X}{A data.table or data.frame whose constant columns we want to remove.}
}
\value{
X_trim A version of X without the constant columns.
}

