% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_na_cols}
\alias{remove_na_cols}
\title{Remove columns of X that are at least some proportion NA}
\usage{
remove_na_cols(X, na_thresh = 1)
}
\arguments{
\item{X}{A data.table or data.frame whose all NA columns we want to remove}

\item{na_thresh}{The proportion of elements in the column that have to be NA
before we remove it. Defaults to 1.}
}

