% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{sum_na}
\alias{sum_na}
\title{Wrapper for sum() that gives NA when all elements are 0}
\usage{
sum_na(x)
}
\description{
We don't want the sums of NAs to be 0.
}
\examples{
sum_na(c(NA, NA, NA))
sum_na(c(0, 1, NA))
sum(c(NA, NA, NA), na.rm = T)
sum(c(0, 1, NA), na.rm = T)
}

